PDF-Comparer/requirements.txt
PyMuPDF==1.23.8  
Pillow==10.2.0   